@@ -11,55 +11,71 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^10.0.14",
-    "@angular/cdk": "^10.1.0",
-    "@angular/common": "^10.0.2",
-    "@angular/compiler": "^10.0.2",
-    "@angular/core": "^10.0.2",
-    "@angular/forms": "~10.0.2",
-    "@angular/platform-browser": "~10.0.2",
-    "@angular/platform-browser-dynamic": "~10.0.2",
-    "@angular/router": "~10.0.2",
+    "@angular/animations": "^16.0.0",
+    "@angular/cdk": "^16.0.0",
+    "@angular/cli": "^16.0.0",
+    "@angular/common": "^16.0.0",
+    "@angular/compiler": "^16.0.0",
+    "@angular/compiler-cli": "^16.0.0",
+    "@angular/core": "^16.0.0",
+    "@angular/forms": "^16.0.0",
+    "@angular/language-service": "^16.0.0",
+    "@angular/material": "^16.2.1",
+    "@angular/platform-browser": "^16.0.0",
+    "@angular/platform-browser-dynamic": "^16.0.0",
+    "@angular/router": "^16.0.0",
+    "@apollo/client": "^3.8.1",
+    "@emailjs/browser": "^3.11.0",
     "@fortawesome/angular-fontawesome": "^0.13.0",
     "@fortawesome/fontawesome-free": "^6.4.2",
     "@fortawesome/fontawesome-svg-core": "^6.4.2",
+    "@fortawesome/free-regular-svg-icons": "^6.4.2",
     "@fortawesome/free-solid-svg-icons": "^6.4.2",
     "angular-colorthief": "^0.1.1",
-    "apollo-angular": "^1.10.0",
-    "apollo-angular-boost": "^1.9.0",
-    "apollo-angular-link-http": "^1.11.0",
-    "apollo-cache-inmemory": "^1.6.0",
-    "apollo-client": "^2.6.0",
-    "apollo-link": "^1.2.14",
-    "apollo-link-context": "^1.0.20",
+    "apollo-angular": "^4.2.0",
     "bootstrap": "^4.5.0",
+    "eslint": "^8.39.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-import": "2.26.0",
+    "eslint-plugin-jsdoc": "^39.6.2",
+    "eslint-plugin-prefer-arrow": "1.2.3",
     "font-awesome": "^4.7.0",
     "github-calendar": "^2.3.2",
-    "graphql": "^15.3.0",
+    "graphql": "^15.8.0",
     "graphql-tag": "^2.10.0",
-    "rxjs": "~6.5.5",
+    "jasmine-core": "~4.5.0",
+    "jasmine-spec-reporter": "~7.0.0",
+    "karma": "~6.4.1",
+    "karma-chrome-launcher": "~3.1.1",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "^2.0.0",
+    "ng-packagr": "^16.0.0",
+    "ngx-lottie": "10.0.0",
+    "prettier": "2.7.1",
+    "protractor": "~7.0.0",
+    "rxjs": "^7.5.7",
     "tslib": "^2.0.0",
-    "zone.js": "~0.10.3"
+    "zone.js": "~0.13.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1000.1",
-    "@angular/cli": "~10.0.1",
-    "@angular/compiler-cli": "~10.0.2",
-    "@types/jasmine": "~3.5.0",
+    "@angular-devkit/build-angular": "^16.0.0",
+    "@angular-devkit/core": "^16.0.0",
+    "@angular-devkit/schematics": "^16.0.0",
+    "@angular-eslint/builder": "^16.0.1",
+    "@angular-eslint/eslint-plugin": "^16.0.1",
+    "@angular-eslint/eslint-plugin-template": "^16.0.1",
+    "@angular-eslint/schematics": "16.0.1",
+    "@angular-eslint/template-parser": "^16.0.1",
+    "@types/jasmine": "~4.3.0",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^12.11.1",
-    "apollo-codegen": "^0.20.2",
+    "@types/node": "~18.16.3",
+    "@typescript-eslint/eslint-plugin": "^5.59.2",
+    "@typescript-eslint/parser": "^5.59.2",
     "codelyzer": "^6.0.0-next.1",
-    "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.0.0",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~3.3.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "protractor": "~7.0.0",
-    "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "typescript": "~3.9.5"
+    "karma-coverage-istanbul-reporter": "^3.0.2",
+    "ts-node": "~10.9.1",
+    "tslint": "^6.1.0",
+    "typescript": "~5.0.4"
   }
 }
