@@ -1,4 +1,4 @@
-import {Component, Input, OnInit} from '@angular/core'
+import {Component, Input, OnInit} from '@angular/core';
 
 @Component({
   selector: 'app-github-repo-card',
@@ -7,8 +7,8 @@ import {Component, Input, OnInit} from '@angular/core'
 })
 export class GithubRepoCardComponent implements OnInit {
   // @Input() github: {name:string, description:string, language:string, fork:number, star:number, size:string, link:string};
-  @Input() github:any;
-  @Input() link:any;
+  @Input() github: any;
+  @Input() link: any;
   constructor(private window: Window) { }
 
   ngOnInit(): void {
@@ -16,7 +16,7 @@ export class GithubRepoCardComponent implements OnInit {
   }
 
   onCardClick(){
-    let win = window.open(this.github.node.url, "_blank");
+    const win = window.open(this.github.node.url, '_blank');
     win.focus();
   }
 
