@@ -8,11 +8,11 @@ import { Component, OnInit } from '@angular/core';
 export class SocialMediaComponent implements OnInit {
   socialMediaLinks = {
 
-    github: "https://github.com/sarthakgoenka",
-    linkedin: "https://www.linkedin.com/in/sarthak-agrawal-a41aa3153/",
-    gmail: "sharthak31@gmail.com",
-    instagram : "https://www.instagram.com/sarthak_goenka/?hl=en",
-    facebook: "https://www.facebook.com/sharthak.agrawal"
+    github: 'https://github.com/Judah-Irani',
+    linkedin: 'https://www.linkedin.com/in/judah-irani-933680112',
+    gmail: 'myfamily4jesus@gmail.com',
+    instagram : 'https://www.instagram.com/judah_irani',
+    facebook: 'https://www.facebook.com/judah.irani'
   };
   constructor() { }
 
